+++ packages/editor/js/editor/sidebar-layout/index.ts
@@ -2,6 +2,7 @@ export { default as SidebarDock } from './SidebarDock';
 export { DEFAULT_SIDEBAR_LAYOUT } from './constants';
 export {
 	getDockSections,
+	getSectionDock,
 	getVisibleDockSections,
 	moveSection,
 	dropIndexFromY,